requirements.txt
google-generativeai
SpeechRecognition
so-vits-svc-fork
TTS
PyAudio
simpleaudio~=1.0.4
torch~=2.2.1+cu121
transformers~=4.38.2
gTTS~=2.5.1